shared/idmap: Move remaining loaders to set_load
